Whitelist processes by command instead of pid
In the event where a process is managed by a service like upstart, and dies,
the pid could change. To prevent cleanslate from killing such processes the
process namd + args should be used for whitelisting.

The number of processes with a given name + arg will still remain the same, i.e. if
the saved process list has three instances of top running in the background only
3 instances of top will be allowed to exist on the next cleanslate run.